break out last_ids into its own module
